+++ openwave/xperiments/E_time_dynamics/E_launcher.py
@@ -6,7 +6,7 @@ XPERIMENT: [WIP] TIME DYNAMICS EXPERIMENT: Time as a function of medium disturba
 import sys
 
 print("Work in Progress XPERIMENT")
-print("- Time as a function of medium disturbance")
+print("- Time as a non-linear function of medium disturbance")
 print("- Time Dynamics")
 
 print("\nWanna join the Time Dynamics Xperiment? Contact us at:")