+++ openwave/xperiments/_archives/spring_mass/_medium.py
@@ -1,7 +1,7 @@
 """
 WAVE-MEDIUM
 
-GRANULE-BASED MEDIUM
+GRANULE-MOTION MEDIUM
 
 Objects Engine @spacetime module.